refactor(discover-daily): move pagination bars to dedicated component
- Extracts PaginationBars into its own reusable component file
- Positions pagination bars at top of details section with proper spacing
- Adds 20px top padding to pagination bars
- Adjusts spacing between bars for better appearance
- Removes image height adjustment since bars are now outside of image container
